0104_update 同步develop分支

+++ src/modules/apa_function/apa_common/apa_context.cc
@@ -116,7 +116,7 @@ const std::string GetRePlanReasonString(const uint8_t type) {
 
 const int CalProjIndexFromPlanningTraj(
     const iflyauto::TrajectoryPoint *trajectory_points, const int n,
-    const int x, const int y) {
+    const double x, const double y) {
   int index = -1;
   double min_dist = std::numeric_limits<double>::infinity();
   for (int i = 0; i < n; ++i) {
